Restructured Jobs packagestandards as per better code organization standards
