--- UDFs/.idea/workspace.xml
@@ -1,84 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="BranchesTreeState">
-    <expand>
-      <path>
-        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
-        <item name="LOCAL_ROOT" type="e8cecc67:BranchNodeDescriptor" />
-      </path>
-      <path>
-        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
-        <item name="REMOTE_ROOT" type="e8cecc67:BranchNodeDescriptor" />
-      </path>
-      <path>
-        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
-        <item name="REMOTE_ROOT" type="e8cecc67:BranchNodeDescriptor" />
-        <item name="GROUP_NODE:origin" type="e8cecc67:BranchNodeDescriptor" />
-      </path>
-    </expand>
-    <select />
-  </component>
-  <component name="ChangeListManager">
-    <list default="true" id="b25167e6-8320-46f6-962a-aa6678b3f9e5" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/../jobs/daily_sale.py" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/../tests/.ipynb_checkpoints/postman assignment-checkpoint.ipynb" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/../tests/location/._SUCCESS.crc" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/../tests/location/.part-00000-d0606b45-54a2-47bd-9440-815dea795bd4-c000.json.crc" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/../tests/location/_SUCCESS" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/../tests/location/part-00000-d0606b45-54a2-47bd-9440-815dea795bd4-c000.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../README.md" beforeDir="false" afterPath="$PROJECT_DIR$/../README.md" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../derby.log" beforeDir="false" afterPath="$PROJECT_DIR$/../derby.log" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../main.py" beforeDir="false" afterPath="$PROJECT_DIR$/../main.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../packages.zip" beforeDir="false" afterPath="$PROJECT_DIR$/../packages.zip" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../tests/.ipynb_checkpoints/Scratch Notebook-checkpoint.ipynb" beforeDir="false" afterPath="$PROJECT_DIR$/../tests/.ipynb_checkpoints/Scratch Notebook-checkpoint.ipynb" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../tests/.ipynb_checkpoints/tests-checkpoint.ipynb" beforeDir="false" afterPath="$PROJECT_DIR$/../tests/.ipynb_checkpoints/tests-checkpoint.ipynb" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../tests/Scratch Notebook.ipynb" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/../tests/tests.ipynb" beforeDir="false" afterPath="$PROJECT_DIR$/../tests/tests.ipynb" afterDir="false" />
-    </list>
-    <option name="SHOW_DIALOG" value="false" />
-    <option name="HIGHLIGHT_CONFLICTS" value="true" />
-    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
-    <option name="LAST_RESOLUTION" value="IGNORE" />
-  </component>
-  <component name="CodeStyleSettingsInfer">
-    <option name="done" value="true" />
-  </component>
-  <component name="Git.Settings">
-    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/.." />
-  </component>
-  <component name="ProjectCodeStyleSettingsMigration">
-    <option name="version" value="1" />
-  </component>
-  <component name="ProjectId" id="1irti0ErqIDT4AmTFI3Vql88Zus" />
-  <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
-  <component name="ProjectViewState">
-    <option name="hideEmptyMiddlePackages" value="true" />
-    <option name="showLibraryContents" value="true" />
-  </component>
-  <component name="PropertiesComponent">
-    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
-    <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
-    <property name="settings.editor.selected.configurable" value="preferences.general" />
-  </component>
-  <component name="TaskManager">
-    <task active="true" id="Default" summary="Default task">
-      <changelist id="b25167e6-8320-46f6-962a-aa6678b3f9e5" name="Default Changelist" comment="" />
-      <created>1602677662989</created>
-      <option name="number" value="Default" />
-      <option name="presentableId" value="Default" />
-      <updated>1602677662989</updated>
-    </task>
-    <servers />
-  </component>
-  <component name="Vcs.Log.Tabs.Properties">
-    <option name="TAB_STATES">
-      <map>
-        <entry key="MAIN">
-          <value>
-            <State />
-          </value>
-        </entry>
-      </map>
-    </option>
-  </component>
-</project>
