+++ src/jobs/country_color_aggregations.py
