+++ src/jobs/daily_sale.py
