+++ src/jobs/max_quantity_rolling.py
