--- tests/notebooks/location/part-00000-d0606b45-54a2-47bd-9440-815dea795bd4-c000.json
@@ -1,1 +1,0 @@
-{"categoryData":"\n                                         {\"categories\" :\n\n                                            [{\"id\": 1, \n                                            \"name\" : \"cat_1\", \n                                            \"sections\":[\n                                                {\n                                                \"id\": 1,\n                                                \"name\": \"sec_1\",\n                                                \"articles\":[{\n                                                                \"id\":11,\n                                                                \"name\": \"art_11\"\n                                                            },\n                                                            {\n                                                                \"id\":12,\n                                                                \"name\": \"art_12\"\n                                                            },\n                                                            {\n                                                                \"id\":13,\n                                                                \"name\": \"art_13\"\n                                                            }]\n                                                },\n                                                {\n                                                \"id\": 2,\n                                                \"name\": \"sec_2\",\n                                                \"articles\":[{\n                                                                \"id\":21,\n                                                                \"name\": \"art_21\"\n                                                            },\n                                                            {\n                                                                \"id\":22,\n                                                                \"name\": \"art_22\"\n                                                            },\n                                                            {\n                                                                \"id\":23,\n                                                                \"name\": \"art_23\"\n                                                            }]\n                                                }]\n                                                },\n                                            {\"id\": 2,\n                                            \"name\" : \"cat_2\", \n                                            \"sections\":[\n                                                {\n                                                \"id\": 3,\n                                                \"name\": \"sec_3\",\n                                                \"articles\":[{\n                                                                \"id\":31,\n                                                                \"name\": \"art_31\"\n                                                            },\n                                                            {\n                                                                \"id\":32,\n                                                                \"name\": \"art_32\"\n                                                            },\n                                                            {\n                                                                \"id\":33,\n                                                                \"name\": \"art_33\"\n                                                            }]\n                                                },\n                                                {\n                                                \"id\": 4,\n                                                \"name\": \"sec_4\",\n                                                \"articles\":[{\n                                                                \"id\":41,\n                                                                \"name\": \"art_41\"\n                                                            },\n                                                            {\n                                                                \"id\":42,\n                                                                \"name\": \"art_42\"\n                                                            },\n                                                            {\n                                                                \"id\":43,\n                                                                \"name\": \"art_43\"\n                                                            }]\n                                                }]\n                                                }\n                                                ]\n                                                }"}
