+++ tests/unit/test_country_color_aggregations.py
@@ -1,6 +1,6 @@
 import pytest
 
-from jobs import country_color_aggregations
+from src.jobs import country_color_aggregations
 from dependencies import spark
