--- tests/unit_tests/test_fixtures/conftest.py
@@ -1,6 +1,0 @@
-from dependencies import spark
-
-import pytest
-
-
-
